finish moving merge -> match
